// src/app/shared/interfaces/user.interface.ts
/*
============================================
interface file
;===========================================
*/

export interface User {
  _id?: string;
  userName?: string;
  password?: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  address: string;
  email: string;
  role: string;
}